extract out trf
